refactor: rename renderer files
